move AppLocator types to discover-utils
